major refactoring of JNI code, API unaffected
The JNI code is now factored in such a way that the functionality of the Java
bindings can easily be compiled into a C library for general consumption.

+++ src/org/puredata/core/PdBase.java
@@ -37,7 +37,7 @@ import java.util.Map;
 
 public final class PdBase {
 
-	private final static Map<String, Long> bindings = new HashMap<String, Long>();
+	private final static Map<String, Integer> bindings = new HashMap<String, Integer>();
 	
 	static {
 		System.loadLibrary("pdnative");
@@ -56,7 +56,7 @@ public final class PdBase {
 	 */
 	public synchronized static void release() {
 		setReceiver(null);
-		for (Long ptr: bindings.values()) {
+		for (int ptr: bindings.values()) {
 			unbindSymbol(ptr);
 		}
 		bindings.clear();
@@ -125,7 +125,7 @@ public final class PdBase {
 		if (bindings.get(symbol) != null) {
 			return 0;
 		}
-		long ptr = bindSymbol(symbol);
+		int ptr = bindSymbol(symbol);
 		if (ptr == 0) {
 			return -1;
 		}
@@ -140,7 +140,7 @@ public final class PdBase {
 	 * @return error code, 0 on success
 	 */
 	public synchronized static int unsubscribe(String symbol) {
-		Long ptr = bindings.get(symbol);
+		Integer ptr = bindings.get(symbol);
 		if (ptr == null) {
 			return 0;
 		}
@@ -229,6 +229,6 @@ public final class PdBase {
 	private native static void addSymbol(String s);
 	private native static int finishList(String receive);
 	private native static int finishMessage(String receive, String message);
-	private native static long bindSymbol(String s);
-	private native static int unbindSymbol(long p);
+	private native static int bindSymbol(String s);
+	private native static int unbindSymbol(int p);
 }
